+++ src/graph_pes/scripts/train.py
@@ -187,6 +187,7 @@ Output for this training run can be found at:
         model,
         test_datasets,
         loader_kwargs=config.fitting.loader_kwargs,
+        logging_prefix="best_model",
         user_eval_metrics=[],
     )