+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 Added support for using arbitrary ``mace-torch`` models within ``graph-pes``, including the ``MACE-MP`` and ``MACE-OFF`` foundation models.
 
+Support for custom batchers for properties in the ``other`` field via the `@register_custom_batcher` decorator.
+
 ## [0.0.17] - 2024-12-19
 
 Add support for ASE `.db` files for backing datasets.