+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.0.18] - 2025-01-03
+
 ### Added
 
 Added support for using arbitrary ``mace-torch`` models within ``graph-pes``, including the ``MACE-MP`` and ``MACE-OFF`` foundation models.